backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.8.1
python-dotenv==1.0.0
